refactor(apps/spu-ui):  organize plan schemas and forms under `_components/plans` and identify them with a setup prefix
