+++ apps/spu-ui/src/app/_components/plans/schemas/setup1-clean-and-acquire.ts
