+++ apps/spu-ui/src/app/_components/plans/schemas/setup1-complete-acquisition.ts
@@ -7,7 +7,7 @@ import {
   trayColumns,
   trayOptions,
   trayRows,
-} from "../../../app/_components/store/setup1/constants";
+} from "../../store/setup1/constants";
 
 const name = "setup1_complete_standard_acquisition";
