+++ apps/spu-ui/src/app/_components/plans/schemas/setup1-standard-cleaning-procedure.ts
